Mainly Refactoring - Click for more info
