--- Dadabase/classes/Frost_Player.py
@@ -1,5 +1,0 @@
-class Frost_Player:
-    def __init__(self, name, level, xp):
-        self.name = name
-        self.level = level
-        self.xp = xp
