+++ Dadabase/data/servers/705783420189671458.json
@@ -1,1 +1,1 @@
-{"name": "Skyward", "links": [{"brawlhalla_id": 7364605, "brawlhalla_name": "", "discord_id": 413070742591373314, "discord_name": "crossychainsaw"}, {"brawlhalla_id": 4, "brawlhalla_name": "", "discord_id": "4", "discord_name": "f"}, {"brawlhalla_id": 5, "brawlhalla_name": "", "discord_id": "5", "discord_name": "5"}, {"brawlhalla_id": 7, "brawlhalla_name": "", "discord_id": "6", "discord_name": "6"}]}
+{"name": "Skyward", "title": "Leaderboard", "links": [{"brawlhalla_id": 7364605, "brawlhalla_name": "", "discord_id": 413070742591373314, "discord_name": "crossychainsaw"}, {"brawlhalla_id": 4, "brawlhalla_name": "", "discord_id": "4", "discord_name": "f"}, {"brawlhalla_id": 5, "brawlhalla_name": "", "discord_id": "5", "discord_name": "5"}, {"brawlhalla_id": 7, "brawlhalla_name": "", "discord_id": "6", "discord_name": "6"}]}
