+++ Dadabase/modules/data.py
@@ -1,5 +1,6 @@
 import json
 
+DATA_LINKS_LOCATION_SERVER_SINGLE_ID = 'Dadabase/data/servers/'
 
 def read_link_data(path, id):
   with open(path + str(id) + '.json') as data:
