--- Dadabase/modules/frost_commands/legend_help.py
@@ -1,64 +1,0 @@
-import discord
-
-async def legend_help(ctx):
-    embed = discord.Embed(title="Legend IDs", description="")
-    embed.description += "Bodvar - 3\n"
-    embed.description += "Cassidy - 4\n"
-    embed.description += "Orion - 5\n"
-    embed.description += "Lord Vraxx - 6\n"
-    embed.description += "Gnash - 7\n"
-    embed.description += "Queen Nai - 8\n"
-    embed.description += "Hattori - 10\n"
-    embed.description += "Sir Roland - 11\n"
-    embed.description += "Scarlet - 12\n"
-    embed.description += "Thatch - 13\n"
-    embed.description += "Ada - 14\n"
-    embed.description += "Sentinel - 15\n"
-    embed.description += "Lucien - 9\n"
-    embed.description += "Teros - 16\n"
-    embed.description += "Brynn - 19\n"
-    embed.description += "Asuri - 20\n"
-    embed.description += "Barraza - 21\n"
-    embed.description += "Ember - 18\n"
-    embed.description += "Azoth - 23\n"
-    embed.description += "Koji - 24\n"
-    embed.description += "Ulgrim - 22\n"
-    embed.description += "Diana - 25\n"
-    embed.description += "Jhala - 26\n"
-    embed.description += "Kor - 28\n"
-    embed.description += "Wu Shang - 29\n"
-    embed.description += "Val - 30\n"
-    embed.description += "Ragnir - 31\n"
-    embed.description += "Cross - 32\n"
-    embed.description += "Mirage - 33\n"
-    embed.description += "Nix - 34\n"
-    embed.description += "Mordex - 35\n"
-    embed.description += "Yumiko - 36\n"
-    embed.description += "Artemis - 37\n"
-    embed.description += "Caspian - 38\n"
-    embed.description += "Sidra - 39\n"
-    embed.description += "Xull - 40\n"
-    embed.description += "Kaya - 42\n"
-    embed.description += "Isaiah - 41\n"
-    embed.description += "Jiro - 43\n"
-    embed.description += "Lin Fei - 44\n"
-    embed.description += "Zariel - 45\n"
-    embed.description += "Rayman - 46\n"
-    embed.description += "Dusk - 47\n"
-    embed.description += "Fait - 48\n"
-    embed.description += "Thor - 49\n"
-    embed.description += "Petra - 50\n"
-    embed.description += "Vector - 51\n"
-    embed.description += "Volkov - 52\n"
-    embed.description += "Onyx - 53\n"
-    embed.description += "Jaeyun - 54\n"
-    embed.description += "Mako - 55\n"
-    embed.description += "Magyar - 56\n"
-    embed.description += "Reno - 57\n"
-    embed.description += "Munin - 58\n"
-    embed.description += "Arcadia - 59\n"
-    embed.description += "Ezio - 60\n"
-    embed.description += "Tezca - 63\n"
-    embed.description += "Thea - 62\n"
-    embed.description += "RedRaptor - 17\n" 
-    await ctx.channel.send(embed=embed)
