+++ Ranknir/data/turn_switch.json
@@ -1,1 +1,1 @@
-{"turn": 4}
+{"turn": 7}
