+++ Ranknir/modules/embed.py
@@ -42,7 +42,7 @@ def prepare_embeds_server(server, players_sorted):
         title=server.get_server_title(), description='', color=server.color)
     if server.member_count == 'show':
         embed_title = __add_member_count(
-            [{"clan": []}], embed_title, len(players_sorted))
+            [{"clan": []}], embed_title, 0, players_sorted)
     # Variables
     embed_array = []
     global rank
